backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# LLM and AI - Let pip resolve compatible versions
openai==2.14.0
anthropic==0.75.0
langchain==1.2.3
langchain-openai==1.1.7
langchain-anthropic==1.3.1
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-classic
langchain-huggingface==1.2.0
langchain-core==1.2.7
tiktoken
ddgs
langchain_text_splitters

# Vector Store
chromadb==1.4.0
sentence-transformers==3.4.1

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
python-pptx==1.0.2
pandas==2.2.3
openpyxl==3.1.5
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
aiofiles==24.1.0
pydantic==2.10.4
pydantic-settings==2.10.1
httpx==0.28.1
tenacity==9.0.0
redis==5.2.1
email_validator

# Code analysis (for GitHub integration)
gitpython==3.1.43

# CORS
fastapi-cors==0.0.6

# BigQuery Support
sqlalchemy-bigquery==1.16.0
google-cloud-bigquery==3.38.0
google-auth==2.41.1

# ETL
pyspark==3.5.3
google-cloud-bigquery-storage==2.36.0
networkx==3.5
